fix: Edam Files format is incompatible with TerosHDL's file format
Edam's 'File' field must contain the 'file_type' in a format like
'vhdlSource-2008'.

Latest TerosHDL's project structure changes have a different approach
for version storage. It is done by storing a version information in
a separate 'file_version' field, while 'file_type' still used as
a file type.

For example: 'file_type' is 'vhdlSource' and 'file_version' is '2008'.
So that change pass thefile information structure to Edalize in a wrong
format. That is cause of loosing version information.

To fix that, a file type and a file version must be combined back.